rag/requirements.txt
# RAG System Dependencies
# Python 3.11+

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==5.1.0

# Embeddings and NLP (using OpenRouter via OpenAI SDK)
openai==1.12.0
tiktoken==0.5.2

# Vector database
qdrant-client==1.7.0

# Google Drive API
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1

# Text processing
nltk==3.8.1
